Introduced MongoDB functionality. Moved CRDB/Mongo into new namespaces.
- Created a new connection string for MongoDB
- target type "query" classes moved under Cockroach.Database
- target type "Dto" classes moved under Cockroach.TargetTypes
- Created a Mongo.TargetTypes namespace
- MongoTargetController to handle CRUD operations for MongoDB backend
