+++ poc/cpat-core/DataAccess/DataControl/Cockroach/DeviceQuery.cs
@@ -1,7 +1,6 @@
-﻿using cpat_core.DataAccess.TargetTypes;
+﻿using cpat_core.DataAccess.DataTransferModels.Cockroach.TargetTypes;
 using cpat_core.Models;
 using Npgsql;
-using NPoco;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,46 +8,46 @@ using System.Threading.Tasks;
 
 namespace cpat_core.DataAccess.DataControl
 {
-    public class TargetQuery
+    public class DeviceQuery
     {
-        private Database dbAccess;
-        private const string _TABLE_NAME_ = "target";
+        public Database dbAccess;
+        private const string _TABLE_NAME_ = "device";
 
-        public TargetQuery()
+        public DeviceQuery()
         {
             dbAccess = new DataAccess.Database();
         }
 
         /// <summary>
-        /// Retrieve a single <c>Target</c> from the database using <c>TargetDto</c> as a mediator.
+        /// Retrieve a single <c>Device</c>.
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public Target GetSingle(Guid id)
+        public Device GetSingle(Guid id)
         {
-            Target target;
+            Device device;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    target = Target.Translate( 
-                        db.Fetch<TargetDto>(new NPoco.Sql(
+                    device = Device.Translate(
+                        db.Fetch<DeviceDto>(new NPoco.Sql(
                             $@"
                                 select * 
                                 from {_TABLE_NAME_}
                                 where id  = @0
-                            ", id)).FirstOrDefault() 
+                            ", id)).FirstOrDefault()
                     );
                     db.Connection.Close();
                 }
             }
 
-            return target;
+            return device;
         }
 
-        public IEnumerable<Target> GetCollection(IEnumerable<Guid> ids)
+        public IEnumerable<Device> GetCollection(IEnumerable<Guid> ids)
         {
             //using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             //{
@@ -62,20 +61,20 @@ namespace cpat_core.DataAccess.DataControl
             //    }
             //}
 
-            return new List<Target>();
+            return new List<Device>();
         }
 
         /// <summary>
-        /// Retrieve a lsit of <c>Target</c> one page at a time. Change the value of page and pageSize to alter the amount of
-        /// <c>Target</c> returned on each query.
+        /// Retrieve a lsit of <c>Device</c> one page at a time. Change the value of page and pageSize to alter the amount of
+        /// <c>Device</c> returned on each query.
         /// </summary>
         /// <param name="page"></param>
         /// <param name="pageSize"></param>
         /// <param name="minval"></param>
         /// <returns></returns>
-        public IEnumerable<Target> GetPage(int page, int pageSize, DateTime minval)
+        public IEnumerable<Device> GetPage(int page, int pageSize, DateTime minval)
         {
-            var list = new List<Target>();
+            var list = new List<Device>();
             minval = minval == null ? new DateTime(year: 2000, month: 0, day: 0) : minval;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -86,8 +85,8 @@ namespace cpat_core.DataAccess.DataControl
 
                     // NOTE: This hard-coding a minimum date value is a temporary fix. Later on, we
                     //      want to change this to some minimum value of the TIMESTAMP data type.
-                    list = Target.Translate(
-                        db.Fetch<TargetDto>(new NPoco.Sql(
+                    list = Device.Translate(
+                        db.Fetch<DeviceDto>(new NPoco.Sql(
                             $@"select * 
                                 from {_TABLE_NAME_}
                                 where datecreated > TIMESTAMP '2000-01-01'
@@ -103,50 +102,46 @@ namespace cpat_core.DataAccess.DataControl
             return list;
         }
 
-
         /// <summary>
-        /// Insert a single <c>Target</c> object.
+        /// Insert a single <c>Device</c> object.
         /// </summary>
-        /// <param name="target"></param>
-        public void Insert(Target target)
-        {
-            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
-            {
-                using (var db = new NPoco.Database(conn))
-                {
-                    var dto = TargetDto.Translate(target);
-
-                    db.Connection.Open();
-                    db.Save<TargetDto>(TargetDto.Translate(target));
-                    db.Connection.Close();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Batch inserts a collection of <c>Target</c> objects.
-        /// </summary>
-        /// <param name="targets"></param>
-        public void Insert(IEnumerable<Target> targets)
+        /// <param name="device"></param>
+        public void Insert(Device device)
         {
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    db.InsertBatch<TargetDto>(TargetDto.Translate(targets.ToList()));
+                    db.Save<DeviceDto>(DeviceDto.Translate(device));
                     db.Connection.Close();
                 }
             }
         }
 
         /// <summary>
-        /// Update a <c>TargetDto</c> object.
+        /// Batch inserts a collection of <c>Device</c> objects.
         /// </summary>
-        /// <param name="docId"></param>
-        /// <param name="target">A <c>Target</c>. This will be translated to a <c>TargetDto</c> before the update operation.</param>
+        /// <param name="devices"></param>
+        public void Insert(IEnumerable<Device> devices)
+        {
+            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
+            {
+                using (var db = new NPoco.Database(conn))
+                {
+                    db.Connection.Open();
+                    db.InsertBatch<DeviceDto>(DeviceDto.Translate(devices.ToList()));
+                    db.Connection.Close();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Update a <c>DeviceDto</c> object.
+        /// </summary>
+        /// <param name="device">A <c>Device</c>. This will be translated to a <c>DeviceDto</c> before the update operation.</param>
         /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Update(Guid docId, Target target)
+        public int Update(Device device)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -154,7 +149,7 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Update(TargetDto.Translate(target), docId);
+                    res = db.Update(DeviceDto.Translate(device));
                     db.Connection.Close();
                 }
             }
@@ -162,15 +157,14 @@ namespace cpat_core.DataAccess.DataControl
             return res;
         }
 
-
         /// <summary>
-        /// Partial update for <c>TargetDto</c> by only updating the specified columns.
+        /// Partial update for <c>DeviceDto</c> by only updating the specified columns.
         /// </summary>
         /// <param name="docId"></param>
-        /// <param name="target"></param>
+        /// <param name="device"></param>
         /// <param name="ops">A <c>string</c> collection </param>
         /// <returns></returns>
-        public int PartialUpdate(Guid docId, Target target, IEnumerable<string> ops)
+        public int PartialUpdate(Guid docId, Device device, IEnumerable<string> ops)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -178,7 +172,7 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Update(TargetDto.Translate(target), docId, ops);
+                    res = db.Update(DeviceDto.Translate(device), docId, ops);
                     db.Connection.Close();
                 }
             }
@@ -187,11 +181,11 @@ namespace cpat_core.DataAccess.DataControl
         }
 
         /// <summary>
-        /// Deletes a <c>TargetDto</c> object.
+        /// Deletes a <c>DeviceDto</c> object.
         /// </summary>
-        /// <param name="target">A <c>Target</c>. This will be translated to a <c>TargetDto</c> before the update operation.</param>
+        /// <param name="device">A <c>Device</c>. This will be translated to a <c>DeviceDto</c> before the update operation.</param>
         /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Remove(Target target)
+        public int Remove(Device device)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -199,34 +193,12 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Delete(TargetDto.Translate(target));
+                    res = db.Delete(DeviceDto.Translate(device));
                     db.Connection.Close();
                 }
             }
 
             return res;
-        }
-
-        /// <summary>
-        /// Updates a <c>TargetDto</c> object as the configured "chosen"
-        /// </summary>
-        /// <param name="target"></param>
-        /// <returns></returns>
-        public int SetTarget(string target)
-        {
-            int res;
-            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
-            {
-                using (var db = new NPoco.Database(conn))
-                {
-                    db.Connection.Open();
-                    //res = db.Update(TargetDto.Translate(target));
-                    db.Connection.Close();
-                }
-            }
-
-            //return res;
-            return 0;
         }
     }
 }
