+++ poc/cpat-core/DataAccess/DataControl/Cockroach/PersonQuery.cs
@@ -1,52 +1,49 @@
-﻿using cpat_core.DataAccess.TargetTypes;
+﻿using cpat_core.DataAccess.DataTransferModels.Cockroach.TargetTypes;
 using cpat_core.Models;
 using Npgsql;
+using NPoco;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace cpat_core.DataAccess.DataControl
 {
-    public class CompanyQuery
+    public class PersonQuery
     {
-        private Database dbAccess;
-        private const string _TABLE_NAME_ = "company";
+        public Database dbAccess;
+        private const string _TABLE_NAME_ = "person";
 
-        public CompanyQuery()
+        public PersonQuery()
         {
             dbAccess = new DataAccess.Database();
         }
 
-        /// <summary>
-        /// Retrieve a single <c>Company</c>.
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        public Company GetSingle(Guid id)
+        public Person GetSingle(Guid id)
         {
-            Company company;
+            Person person;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    company = Company.Translate(
-                        db.Fetch<CompanyDto>(new NPoco.Sql(
+                    person = Person.Translate(
+                        db.Fetch<PersonDto>(new Sql(
                             $@"
-                                select * 
-                                from {_TABLE_NAME_}
-                                where id  = @0
-                            ", id)).FirstOrDefault()
+                            select * 
+                            from {_TABLE_NAME_}
+                            where id  = @0
+                            ", id)).FirstOrDefault()    
                     );
                     db.Connection.Close();
                 }
             }
 
-            return company;
+            return person;
         }
 
-        public IEnumerable<Company> GetCollection(IEnumerable<Guid> ids)
+        public IEnumerable<Person> GetCollection(IEnumerable<Guid> ids)
         {
             //using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             //{
@@ -60,20 +57,20 @@ namespace cpat_core.DataAccess.DataControl
             //    }
             //}
 
-            return new List<Company>();
+            return new List<Person>();
         }
 
         /// <summary>
-        /// Retrieve a lsit of <c>Company</c> one page at a time. Change the value of page and pageSize to alter the amount of
-        /// <c>Company</c> returned on each query.
+        /// Retrieve a lsit of <c>Person</c> one page at a time. Change the value of page and pageSize to alter the amount of
+        /// <c>Target</c> returned on each query.
         /// </summary>
         /// <param name="page"></param>
         /// <param name="pageSize"></param>
         /// <param name="minval"></param>
         /// <returns></returns>
-        public IEnumerable<Company> GetPage(int page, int pageSize, DateTime minval)
+        public IEnumerable<Person> GetPage(int page, int pageSize, DateTime minval)
         {
-            var list = new List<Company>();
+            var list = new List<Person>();
             minval = minval == null ? new DateTime(year: 2000, month: 0, day: 0) : minval;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -84,8 +81,8 @@ namespace cpat_core.DataAccess.DataControl
 
                     // NOTE: This hard-coding a minimum date value is a temporary fix. Later on, we
                     //      want to change this to some minimum value of the TIMESTAMP data type.
-                    list = Company.Translate(
-                        db.Fetch<CompanyDto>(new NPoco.Sql(
+                    list = Person.Translate(
+                        db.Fetch<PersonDto>(new NPoco.Sql(
                             $@"select * 
                                 from {_TABLE_NAME_}
                                 where datecreated > TIMESTAMP '2000-01-01'
@@ -102,54 +99,57 @@ namespace cpat_core.DataAccess.DataControl
         }
 
         /// <summary>
-        /// Insert a single <c>Company</c> object.
+        /// Insert a single <c>Person</c> object.
         /// </summary>
-        /// <param name="company"></param>
-        public void Insert(Company company)
+        /// <param name="person"></param>
+        public void Insert(Person person)
         {
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    db.Save<CompanyDto>(CompanyDto.Translate(company));
+                    
+                    PersonDto pdto = PersonDto.Translate(person);
+                    pdto.DateCreated = DateTime.Now;
+
+                    db.Save<PersonDto>(pdto);
                     db.Connection.Close();
                 }
             }
         }
 
         /// <summary>
-        /// Batch inserts a collection of <c>Company</c> objects.
+        /// Batch inserts a collection of <c>Person</c> objects.
         /// </summary>
-        /// <param name="companies"></param>
-        public void Insert(IEnumerable<Company> companies)
+        /// <param name="devices"></param>
+        public void Insert(IEnumerable<Person> devices)
         {
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    db.InsertBatch<CompanyDto>(CompanyDto.Translate(companies.ToList()));
+                    db.InsertBatch<PersonDto>(PersonDto.Translate(devices.ToList()));
                     db.Connection.Close();
                 }
             }
         }
 
         /// <summary>
-        /// Update a <c>CompanyDto</c> object.
+        /// Update a <c>PersonDto</c> object.
         /// </summary>
-        /// <param name="company">A <c>Comapny</c>. This will be translated to a <c>ComapnyDto</c> before the update operation.</param>
+        /// <param name="device">A <c>Person</c>. This will be translated to a <c>PersonDto</c> before the update operation.</param>
         /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Update(Company company)
+        public int Update(Person device)
         {
             int res;
-
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Update(CompanyDto.Translate(company));
+                    res = db.Update(PersonDto.Translate(device));
                     db.Connection.Close();
                 }
             }
@@ -158,13 +158,13 @@ namespace cpat_core.DataAccess.DataControl
         }
 
         /// <summary>
-        /// Partial update for <c>CompanyDto</c> by only updating the specified columns.
+        /// Partial update for <c>PersonDto</c> by only updating the specified columns.
         /// </summary>
         /// <param name="docId"></param>
-        /// <param name="company"></param>
+        /// <param name="person"></param>
         /// <param name="ops">A <c>string</c> collection </param>
         /// <returns></returns>
-        public int PartialUpdate(Guid docId, Company company, IEnumerable<string> ops)
+        public int PartialUpdate(Guid docId, Person person, IEnumerable<string> ops)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -172,7 +172,7 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Update(CompanyDto.Translate(company), docId, ops);
+                    res = db.Update(PersonDto.Translate(person), docId, ops);
                     db.Connection.Close();
                 }
             }
@@ -181,11 +181,11 @@ namespace cpat_core.DataAccess.DataControl
         }
 
         /// <summary>
-        /// Deletes a <c>CompanyDto</c> object.
+        /// Deletes a <c>PersonDto</c> object.
         /// </summary>
-        /// <param name="company">A <c>Comapny</c>. This will be translated to a <c>ComapnyDto</c> before the update operation.</param>
+        /// <param name="device">A <c>Person</c>. This will be translated to a <c>PersonDto</c> before the update operation.</param>
         /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Remove(Company company)
+        public int Remove(Person device)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -193,7 +193,7 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Delete(CompanyDto.Translate(company));
+                    res = db.Delete(PersonDto.Translate(device));
                     db.Connection.Close();
                 }
             }
