+++ poc/cpat-core/DataAccess/DataControl/Cockroach/TargetQuery.cs
@@ -1,6 +1,7 @@
-﻿using cpat_core.DataAccess.TargetTypes;
+﻿using cpat_core.DataAccess.DataTransferModels.Cockroach.TargetTypes;
 using cpat_core.Models;
 using Npgsql;
+using NPoco;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,46 +9,46 @@ using System.Threading.Tasks;
 
 namespace cpat_core.DataAccess.DataControl
 {
-    public class DeviceQuery
+    public class TargetQuery
     {
-        public Database dbAccess;
-        private const string _TABLE_NAME_ = "device";
+        private Database dbAccess;
+        private const string _TABLE_NAME_ = "target";
 
-        public DeviceQuery()
+        public TargetQuery()
         {
             dbAccess = new DataAccess.Database();
         }
 
         /// <summary>
-        /// Retrieve a single <c>Device</c>.
+        /// Retrieve a single <c>Target</c> from the database using <c>TargetDto</c> as a mediator.
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public Device GetSingle(Guid id)
+        public Target GetSingle(Guid id)
         {
-            Device device;
+            Target target;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    device = Device.Translate(
-                        db.Fetch<DeviceDto>(new NPoco.Sql(
+                    target = Target.Translate( 
+                        db.Fetch<TargetDto>(new NPoco.Sql(
                             $@"
                                 select * 
                                 from {_TABLE_NAME_}
                                 where id  = @0
-                            ", id)).FirstOrDefault()
+                            ", id)).FirstOrDefault() 
                     );
                     db.Connection.Close();
                 }
             }
 
-            return device;
+            return target;
         }
 
-        public IEnumerable<Device> GetCollection(IEnumerable<Guid> ids)
+        public IEnumerable<Target> GetCollection(IEnumerable<Guid> ids)
         {
             //using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             //{
@@ -61,20 +62,20 @@ namespace cpat_core.DataAccess.DataControl
             //    }
             //}
 
-            return new List<Device>();
+            return new List<Target>();
         }
 
         /// <summary>
-        /// Retrieve a lsit of <c>Device</c> one page at a time. Change the value of page and pageSize to alter the amount of
-        /// <c>Device</c> returned on each query.
+        /// Retrieve a lsit of <c>Target</c> one page at a time. Change the value of page and pageSize to alter the amount of
+        /// <c>Target</c> returned on each query.
         /// </summary>
         /// <param name="page"></param>
         /// <param name="pageSize"></param>
         /// <param name="minval"></param>
         /// <returns></returns>
-        public IEnumerable<Device> GetPage(int page, int pageSize, DateTime minval)
+        public IEnumerable<Target> GetPage(int page, int pageSize, DateTime minval)
         {
-            var list = new List<Device>();
+            var list = new List<Target>();
             minval = minval == null ? new DateTime(year: 2000, month: 0, day: 0) : minval;
 
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -85,8 +86,8 @@ namespace cpat_core.DataAccess.DataControl
 
                     // NOTE: This hard-coding a minimum date value is a temporary fix. Later on, we
                     //      want to change this to some minimum value of the TIMESTAMP data type.
-                    list = Device.Translate(
-                        db.Fetch<DeviceDto>(new NPoco.Sql(
+                    list = Target.Translate(
+                        db.Fetch<TargetDto>(new NPoco.Sql(
                             $@"select * 
                                 from {_TABLE_NAME_}
                                 where datecreated > TIMESTAMP '2000-01-01'
@@ -102,69 +103,50 @@ namespace cpat_core.DataAccess.DataControl
             return list;
         }
 
+
         /// <summary>
-        /// Insert a single <c>Device</c> object.
+        /// Insert a single <c>Target</c> object.
         /// </summary>
-        /// <param name="device"></param>
-        public void Insert(Device device)
+        /// <param name="target"></param>
+        public void Insert(Target target)
         {
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
+                    var dto = TargetDto.Translate(target);
+
                     db.Connection.Open();
-                    db.Save<DeviceDto>(DeviceDto.Translate(device));
+                    db.Save<TargetDto>(TargetDto.Translate(target));
                     db.Connection.Close();
                 }
             }
         }
 
         /// <summary>
-        /// Batch inserts a collection of <c>Device</c> objects.
+        /// Batch inserts a collection of <c>Target</c> objects.
         /// </summary>
-        /// <param name="devices"></param>
-        public void Insert(IEnumerable<Device> devices)
+        /// <param name="targets"></param>
+        public void Insert(IEnumerable<Target> targets)
         {
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
             {
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    db.InsertBatch<DeviceDto>(DeviceDto.Translate(devices.ToList()));
+                    db.InsertBatch<TargetDto>(TargetDto.Translate(targets.ToList()));
                     db.Connection.Close();
                 }
             }
         }
 
         /// <summary>
-        /// Update a <c>DeviceDto</c> object.
-        /// </summary>
-        /// <param name="device">A <c>Device</c>. This will be translated to a <c>DeviceDto</c> before the update operation.</param>
-        /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Update(Device device)
-        {
-            int res;
-            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
-            {
-                using (var db = new NPoco.Database(conn))
-                {
-                    db.Connection.Open();
-                    res = db.Update(DeviceDto.Translate(device));
-                    db.Connection.Close();
-                }
-            }
-
-            return res;
-        }
-
-        /// <summary>
-        /// Partial update for <c>DeviceDto</c> by only updating the specified columns.
+        /// Update a <c>TargetDto</c> object.
         /// </summary>
         /// <param name="docId"></param>
-        /// <param name="device"></param>
-        /// <param name="ops">A <c>string</c> collection </param>
-        /// <returns></returns>
-        public int PartialUpdate(Guid docId, Device device, IEnumerable<string> ops)
+        /// <param name="target">A <c>Target</c>. This will be translated to a <c>TargetDto</c> before the update operation.</param>
+        /// <returns><c>int</c> denoting the success value of the operation.</returns>
+        public int Update(Guid docId, Target target)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -172,7 +154,31 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Update(DeviceDto.Translate(device), docId, ops);
+                    res = db.Update(TargetDto.Translate(target), docId);
+                    db.Connection.Close();
+                }
+            }
+
+            return res;
+        }
+
+
+        /// <summary>
+        /// Partial update for <c>TargetDto</c> by only updating the specified columns.
+        /// </summary>
+        /// <param name="docId"></param>
+        /// <param name="target"></param>
+        /// <param name="ops">A <c>string</c> collection </param>
+        /// <returns></returns>
+        public int PartialUpdate(Guid docId, Target target, IEnumerable<string> ops)
+        {
+            int res;
+            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
+            {
+                using (var db = new NPoco.Database(conn))
+                {
+                    db.Connection.Open();
+                    res = db.Update(TargetDto.Translate(target), docId, ops);
                     db.Connection.Close();
                 }
             }
@@ -181,11 +187,11 @@ namespace cpat_core.DataAccess.DataControl
         }
 
         /// <summary>
-        /// Deletes a <c>DeviceDto</c> object.
+        /// Deletes a <c>TargetDto</c> object.
         /// </summary>
-        /// <param name="device">A <c>Device</c>. This will be translated to a <c>DeviceDto</c> before the update operation.</param>
+        /// <param name="target">A <c>Target</c>. This will be translated to a <c>TargetDto</c> before the update operation.</param>
         /// <returns><c>int</c> denoting the success value of the operation.</returns>
-        public int Remove(Device device)
+        public int Remove(Target target)
         {
             int res;
             using (var conn = new NpgsqlConnection(dbAccess.connectionString))
@@ -193,12 +199,34 @@ namespace cpat_core.DataAccess.DataControl
                 using (var db = new NPoco.Database(conn))
                 {
                     db.Connection.Open();
-                    res = db.Delete(DeviceDto.Translate(device));
+                    res = db.Delete(TargetDto.Translate(target));
                     db.Connection.Close();
                 }
             }
 
             return res;
+        }
+
+        /// <summary>
+        /// Updates a <c>TargetDto</c> object as the configured "chosen"
+        /// </summary>
+        /// <param name="target"></param>
+        /// <returns></returns>
+        public int SetTarget(string target)
+        {
+            int res;
+            using (var conn = new NpgsqlConnection(dbAccess.connectionString))
+            {
+                using (var db = new NPoco.Database(conn))
+                {
+                    db.Connection.Open();
+                    //res = db.Update(TargetDto.Translate(target));
+                    db.Connection.Close();
+                }
+            }
+
+            //return res;
+            return 0;
         }
     }
 }
