+++ poc/cpat-core/DataAccess/DataControl/Mongo/Database/Database.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace cpat_core.DataAccess.DataControl.Mongo.Database
+{
+    public class Database
+    {
+        public string connectionString = "";
+
+        public Database()
+        {
+
+        }
+    }
+}
