+++ poc/cpat-core/DataAccess/DataTransferModels/Mongo/TargetTypes/CompanyDto.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace cpat_core.DataAccess.DataTransferModels.Mongo.TargetTypes
+{
+    public class CompanyDto
+    {
+    }
+}
