+++ poc/cpat-core/appsettings.json
@@ -1,4 +1,7 @@
 {
+  "ConnectionStrings": {
+    "CPAT_DATASTORE": "mongodb://192.168.1.44:30001/?readPreference=primary&appname=MongoDB%20Compass&ssl=false"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
